requirements.txt
dspy
openai
datasets
python-dotenv
scikit-learn
